Removed SEGFAULT-inducing Py_DECREFs (see notes).
It turns out PyList_GetItem() /borrows/ references and does not create a
new one. This meant that every Py_DECREF() I loyally appended to the end
of for loops and method calls were causing the objects to be cleaned up
by the GC.

You win some and you lose some.